Added optional Parsedown loading
This change lets the [Parsedown](https://github.com/erusev/parsedown) and the [ParsedownExtra](https://github.com/erusev/parsedown-extra) classes be used to format Markdown instead of the built-in function.

Simply place the *Parsedown.php* and *ParsedownExtra.php* files in the same directory as *index.php*. If only *Parsedown.php* is present, it will use that. Parsedown is optional and Bare will default to the built-in function if these are not present.