Moved from file based cache to SQLite db
Bare processes posts as plain text files with .md extensions, but file based caching was getting too cumbersome. With many pages, the number of subfolders alone an get quite large. This change moves all caching to a SQLite db which will be created on demand

This is a transparent change that shouldn't affect any other functionality.