PHP filemtime() workaround
For some reason, PHP's filemtime() doesn't always give the expected result. Unsure if it's a bug in my code or a PHP problem, but until that's sorted, I'm taking out the post modification check. If a post is modified after it was published, simply delete cache.db and everything will get re-indexed.