Added custom about page
This change adds the option to create multiple about pages in posts/about. The default about page is main.md. Other sub-pages can be created also formatted in Markdown with a .md extension. As with posts, the first line is extracted as the page title and the file name is the URL slug.