Fixed large media skip-head bug
This update fixes a problem when the visitor skips ahead of an embedded video or audio file and the stream not starting/stopping correctly at the requested range position. This also fixes a problem with streaming very large files when the visitor closes the connection mid-playback and memory isn't freed fast enough.